Created components and pages for handling multiple customer contacts.
Component CreateContactComponent is a button with a form
to create a new customer.

Component EditContactComponent is a button with a form to
update contact details.

Component ContactsComponent is a list of contacts.

Customer page has been updated to use the new components.

A component refresher has been programmed to handle changes
in contact lists.